+++ CHANGELOG.md
@@ -1,9 +1,12 @@
 # Version History
 
-## Unreleased
+## 14.0.2
 
+* [THORN-322] Add docs for Health of hardware facing devices.
+* [SKB-1445] Update ``AcquireDataForCalibration`` to be more resilient to DAQ errors.
 * [THORN-125] Improve observability when UNPROGRAMMED.
 * [THORN-125] Bump daq version: 6.1.0 -> 6.1.1
+* [THORN-611] Add `WrenTRL` device property to SpsStation for WREN TRL
 
 ## 14.0.1